backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Google Cloud AI/ML
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.5
google-cloud-firestore==2.13.1
google-generativeai==0.3.2

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.25.2

# Web scraping and data collection
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
